Add files via upload
A bicycle is statically unstable like the inverted pendulum, but under certain conditions, is stable in forward motion. Thus, it is mandatory to create these conditions for a cycle to balance. We have incorporated a Control Moment Gyro-an attitude control device typically used in spacecraft attitude control systems. The bicycle is equipped with sensors for its self-balancing and a system to give it transnational movement.